Isolated mocks, added assertions for new model fields
